Move timeline comp to visualization folder
